+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 1.0.0 (Unreleased)
  - Ability to construct body with content/type [#12](https://github.com/restbeast/restbeast/issues/12)
  - Cookie support [#16](https://github.com/restbeast/restbeast/issues/16)
+ - Apple Silicon M1 support for macOS
 
 0.13.0
 - Request parameters [#13](https://github.com/restbeast/restbeast/issues/13)