+++ CHANGELOG.md
@@ -2,6 +2,7 @@
  - Ability to construct body with content/type [#12](https://github.com/restbeast/restbeast/issues/12)
  - Cookie support [#16](https://github.com/restbeast/restbeast/issues/16)
  - Apple Silicon M1 support for macOS
+ - Enhance json decode error message
 
 0.13.0
 - Request parameters [#13](https://github.com/restbeast/restbeast/issues/13)