+++ CHANGELOG.md
@@ -3,7 +3,8 @@
  - Cookie support [#16](https://github.com/restbeast/restbeast/issues/16)
  - Apple Silicon M1 support for macOS
  - Enhance json decode error message
- - new assertion function `assertNonEmptyString` [#35](https://github.com/restbeast/restbeast/issues/35)
+ - New assertion function `assertNonEmptyString` [#35](https://github.com/restbeast/restbeast/issues/35)
+ - List command [#30](https://github.com/restbeast/restbeast/issues/30)
 
 0.13.0
 - Request parameters [#13](https://github.com/restbeast/restbeast/issues/13)